requirements.txt
# Core ML
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0

# API / Serving
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Experiment Tracking & Versioning (Development)
mlflow==2.8.1
dvc==3.30.1

# Utilities
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
opencv-python==4.8.1.78
pyyaml==6.0.1
